+++ dds/src/implementation/dds/dds_domain_participant.rs
@@ -21,6 +21,7 @@ use crate::{
             discovery_types::{BuiltinEndpointQos, BuiltinEndpointSet},
             endpoint::RtpsEndpoint,
             group::RtpsGroup,
+            messages::overall_structure::RtpsMessageRead,
             participant::RtpsParticipant,
             reader::RtpsReader,
             reader_locator::RtpsReaderLocator,
@@ -42,7 +43,7 @@ use crate::{
                 BUILT_IN_WRITER_WITH_KEY, ENTITYID_PARTICIPANT, USER_DEFINED_READER_GROUP,
                 USER_DEFINED_TOPIC, USER_DEFINED_WRITER_GROUP,
             },
-            writer::RtpsWriter, messages::RtpsMessageRead,
+            writer::RtpsWriter,
         },
         utils::condvar::DdsCondvar,
     },
