+++ dds/src/implementation/rtps/transport.rs
@@ -1,4 +1,4 @@
-use super::{messages::RtpsMessageWrite, types::Locator};
+use super::{messages::overall_structure::RtpsMessageWrite, types::Locator};
 
 pub trait TransportWrite {
     fn write(&mut self, message: &RtpsMessageWrite<'_>, destination_locator_list: &[Locator]);
